backend/functions/sign_upload_url/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.16.0
google-cloud-firestore==2.16.0
firebase-admin==6.5.0
